requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.24.3
torch==2.0.1
roboflow==0.2.27
ultralytics
supervision
opencv-python
matplotlib
requests
tensorflow-cpu==2.13.0
librosa
pydantic==1.10.14
python-dotenv==1.0.1
boto3==1.28.60